highscore refactoring
